show locally store invoices

+++ src/app/modules/my-invoices/my-invoices.component.ts
@@ -16,18 +16,31 @@ export class MyInvoicesComponent implements OnInit {
     this.l_invoicesIds = this.storageService.getItem("l_invoicesIds");
     this.loadInvoices();
   }
-  loadInvoices(){
-    if (this.l_invoicesIds.length > 0) {
+  loadInvoices() {
+    if (this.l_invoicesIds && this.l_invoicesIds.length > 0) {
       for (let id of this.l_invoicesIds) {
         const invoice = this.storageService.getItem(`l_invoice-${id}`);
         if (invoice) this.invoices.push(invoice);
       }
     }
   }
-  deleteInvoice(number: number,index:number) {
-    this.l_invoicesIds.splice(this.l_invoicesIds.indexOf(number), 1);
-    this.storageService.setItem('l_invoicesIds',this.l_invoicesIds);
-    this.storageService.removeItem(`l_invoice-${number}`);
-    this.invoices.splice(index,1);
+  deleteInvoice(number: number, index: number) {
+    if (confirm("Are you sure you want to delete this invoice?")) {
+      this.l_invoicesIds.splice(this.l_invoicesIds.indexOf(number), 1);
+      this.storageService.setItem("l_invoicesIds", this.l_invoicesIds);
+      this.storageService.removeItem(`l_invoice-${number}`);
+      this.invoices.splice(index, 1);
+    }
+  }
+  clearInvoices() {
+    if (
+      confirm(
+        "Are you sure you want to permanently delete all of your invoices?"
+      )
+    ) {
+      this.storageService.clear();
+      this.invoices = [];
+      this.l_invoicesIds = [];
+    }
   }
 }
